Finalized@

+++ DTO/CapitalDTO.cs
@@ -5,7 +5,7 @@ namespace DTO
     public class CapitalDTO
     {
         public string Name { get; set; }
-        public decimal Value { get; set; }
+        public long Value { get; set; }
         public DateTime Date { get; set; }
     }
 }
