+++ DataAccess/Entities/ProfitNLoss.cs
@@ -7,7 +7,7 @@ namespace DataAccess.Entities
     public class ProfitNLoss
     {
         public Guid ID { get; set; }
-        public decimal Value { get; set; }
+        public long Value { get; set; }
         public Strategy Strategy { get; set; }
         public DateTime Date { get; set; }
     }
